Added module token
