+++ src/Results/mod.rs
@@ -1,3 +1,2 @@
 pub mod campus;
 pub mod me;
-pub mod token;
